+++ app/src/main/java/org/oxycblt/auxio/MainActivity.kt
@@ -44,6 +44,10 @@ import org.oxycblt.auxio.util.systemBarInsetsCompat
  *
  * TODO: Add multi-select
  *
+ * TODO: Use proper material attributes (Not the weird dimen attributes I currently have)
+ *
+ * TODO: Test out material animation system
+ *
  * @author OxygenCobalt
  */
 class MainActivity : AppCompatActivity() {