+++ app/src/main/java/org/oxycblt/auxio/music/picker/MusicPickerViewModel.kt
@@ -1,3 +1,20 @@
+/*
+ * Copyright (c) 2022 Auxio Project
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
 package org.oxycblt.auxio.music.picker
 
 import androidx.lifecycle.ViewModel
@@ -13,15 +30,17 @@ class MusicPickerViewModel : ViewModel(), MusicStore.Callback {
     private val musicStore = MusicStore.getInstance()
 
     private val _currentSong = MutableStateFlow<Song?>(null)
-    val currentSong: StateFlow<Song?> get() = _currentSong
+    val currentSong: StateFlow<Song?>
+        get() = _currentSong
 
     private val _currentArtists = MutableStateFlow<List<Artist>?>(null)
-    val currentArtists: StateFlow<List<Artist>?> get() = _currentArtists
+    val currentArtists: StateFlow<List<Artist>?>
+        get() = _currentArtists
 
     fun setSongUid(uid: Music.UID) {
         val library = unlikelyToBeNull(musicStore.library)
         _currentSong.value = library.find(uid)
-        _currentArtists.value =  _currentSong.value?.artists
+        _currentArtists.value = _currentSong.value?.artists
     }
 
     fun setArtistUids(uids: Array<Music.UID>) {
@@ -36,7 +55,7 @@ class MusicPickerViewModel : ViewModel(), MusicStore.Callback {
             if (song != null) {
                 _currentSong.value = library.sanitize(song)
                 _currentArtists.value = _currentSong.value?.artists
-            } else if (artists != null){
+            } else if (artists != null) {
                 _currentArtists.value = artists.mapNotNull { library.sanitize(it) }
             }
         }